refactor(scanner): move secret redaction to redact.py (verbatim)
Pure relocation of _redact_secrets and its regex passes into a new
graphlm/redact.py, re-imported by scanner.py so every existing import
path (incl. tests/test_scanner.py's `from graphlm.scanner import
_redact_secrets`) keeps working. No behaviour change; tests untouched.

Why: scanner.py sat at 637 lines, over the 600-line module limit, and
the next change (tree-sitter skeletons for oversized files, innovation
#2) adds to scan_project. Moving the redaction regexes out is the
cohesive split: they are a standalone text transform with no scanner
state.
